Modify ProfileVC and add FeatureType enum file
